requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
openpyxl
fpdf
